requirements.txt
flask
fastapi
uvicorn[standard]
python-multipart
pydantic
httpx
Pillow
jinja2
python-dotenv
openai
pyttsx3
edge-tts
pydub
ffmpeg-python